Added default list of paths to ignore when scanning for assets
ResourceManager now provides a default array of pruned directory
extensions, which identify directory names which should not be
displayed in resource popups and the resource manager. These include
favorites like .git or .svn, as well as .xcodeproj - resources either
should not exist at these locations, or should not be used by ccb.